requirements.txt
# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0

# YOLO
ultralytics>=8.3.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Annotation / Dataset Tools
pycocotools>=2.0.7

# Geometry
shapely>=2.0.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
requests>=2.31.0
